pico: split display drivers and add "none" driver
A bit more work as there's some shared code here. Default config now has no display driver.
