+++ 32blit-pico/board/pico/config.cmake
@@ -2,6 +2,5 @@ set(BLIT_BOARD_NAME "Pico/default")
 
 # TODO: this is the hacky explorer setup I initially used
 set(BLIT_BOARD_DEFINITIONS
-    DISPLAY_ST7789
     INPUT_GPIO
 )
