+++ 32blit-pico/board/pimoroni_picosystem/config.cmake
@@ -2,7 +2,7 @@ set(BLIT_BOARD_NAME "PicoSystem")
 
 set(BLIT_BOARD_DEFINITIONS
     AUDIO_BEEP_PIN=PICOSYSTEM_AUDIO_PIN
-    DISPLAY_ST7789
+    DISPLAY_ST7789 # code in both the sdk and some games still depends on this
     INPUT_GPIO
     LED_R_PIN=PICOSYSTEM_LED_R_PIN
     LED_G_PIN=PICOSYSTEM_LED_G_PIN
@@ -15,3 +15,4 @@ set(BLIT_BOARD_DEFINITIONS
 )
 
 blit_driver(audio beep)
+blit_driver(display st7789)
